create utils.py to store shared functions between train.py and predict.py
