+++ recognition/vqvae_Carl_Chang/predict.py
@@ -10,6 +10,7 @@ from skimage.metrics import structural_similarity as ssim
 # Import your custom modules
 from modules import VQVAE
 from dataset import get_data_loader 
+from utils import calculate_ssim, show_img, show_combined
 
 # Data path
 test_path = "./data/keras_slices_test"
@@ -19,10 +20,12 @@ batch_size = 256
 num_hiddens = 128
 num_residual_hiddens = 32
 num_residual_layers = 2
-
 embedding_dim = 64
 num_embeddings = 512
 commitment_cost = 0.25
+
+# Prepare Data Loader
+test_loader = get_data_loader(test_path, batch_size=batch_size, norm_image=True)
 
 # Initialize Device
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
@@ -32,9 +35,6 @@ model = VQVAE(num_hiddens, num_residual_layers, num_residual_hiddens, num_embedd
 model.load_state_dict(torch.load("vqvae_model.pth", map_location=device))  # Load the saved model weights
 model.to(device)
 model.eval()  # Set to evaluation mode
-
-# Prepare Data Loader
-test_loader = get_data_loader(test_path, batch_size=batch_size, norm_image=True)
 
 # Perform Inference
 test_originals = next(iter(test_loader)).to(device)
@@ -46,50 +46,17 @@ with torch.no_grad():
 test_reconstructions = test_reconstructions.cpu().data
 test_originals = test_originals.cpu().data
 
-# Calculate SSIM between original and reconstructed images
-def calculate_ssim(original, reconstructed):
-    ssim_scores = []
-    for i in range(original.size(0)):
-        orig = original[i].numpy().squeeze()
-        recon = reconstructed[i].numpy().squeeze()
-        ssim_score = ssim(orig, recon, data_range=recon.max() - recon.min())
-        ssim_scores.append(ssim_score)
-    return np.mean(ssim_scores)
-
 average_ssim = calculate_ssim(test_originals, test_reconstructions)
 print(f"Average SSIM: {average_ssim:.3f}")
 
-# Display Results
-def show(img):
-    npimg = img.numpy()
-    npimg = (npimg - npimg.min()) / (npimg.max() - npimg.min())
-    fig = plt.imshow(np.transpose(npimg, (1, 2, 0)), interpolation='nearest')
-    fig.axes.get_xaxis().set_visible(False)
-    fig.axes.get_yaxis().set_visible(False)
+print("Displaying original images:")
+show_img(make_grid(test_originals[:16], nrow=8, normalize=True))
+plt.show()
 
 print("Displaying reconstructed images:")
-show(make_grid(test_reconstructions[:16], nrow=8, normalize=True))
+show_img(make_grid(test_reconstructions[:16], nrow=8, normalize=True))
 plt.show()
 
-print("Displaying original images:")
-show(make_grid(test_originals[:16], nrow=8, normalize=True))
-plt.show()
-
-# Combined Display of Original and Reconstructed Images
-def show_combined(originals, reconstructions):
-    # Combine originals and reconstructions
-    combined = torch.cat([originals[:8], reconstructions[:8]], dim=0)
-    
-    npimg = make_grid(combined, nrow=8, normalize=True).numpy()
-    npimg = (npimg - npimg.min()) / (npimg.max() - npimg.min())  # Normalize for display
-    
-    # Display with SSIM
-    fig, ax = plt.subplots(figsize=(12, 6))
-    ax.imshow(np.transpose(npimg, (1, 2, 0)), interpolation='nearest')
-    ax.set_title(f"Originals (Top) and Reconstructions (Bottom) - Avg SSIM: {average_ssim:.3f}", fontsize=14)
-    ax.axis('off')
-    plt.show()
-
-# Show Combined Images
-show_combined(test_originals, test_reconstructions)
+print("Displaying original vs reconstructed images:")
+show_combined(test_originals, test_reconstructions, average_ssim)
 plt.show()
